Reorganized namespaces and minor fixes
