+++ src/FeatureManagement.Database.Abstractions/Properties/Assembly.cs
@@ -1,3 +1,6 @@
+// Copyright (c) Matteo Ciapparelli.
+// Licensed under the MIT license.
+
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("FeatureManagement.Database.Tests")]
