+++ tests/FeatureManagement.Database.Tests/FeatureManagement/Database/Tests/CachedFeatureStoreTests.cs
@@ -1,7 +1,6 @@
 // Copyright (c) Matteo Ciapparelli.
 // Licensed under the MIT license.
 
-using FeatureManagement.Database.Abstractions;
 using Microsoft.Extensions.Caching.Distributed;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
